main: use no timeout with reader.wait and move it up
It seems like this fixes (hard to reproduce) missing messages, e.g. the
second MESSAGE shows up in the journal, but gets not into
`client.log_messages`.

> 2018-05-29 01:35:00,685:INFO:i-0d2ea76e3e48e65d8:celery.worker.strategy:[none]: Received task: project.celery.send_password_reset_email_task…
> 2018-05-29 01:35:00,774:INFO:i-0d2ea76e3e48e65d8:project.app.utils:[none]: Sending email to …
> 2018-05-29 01:35:00,817:DEBUG:i-0d2ea76e3e48e65d8:django_ses:[none]: send_messages.throttle rate_limit='14.0'"…
> 2018-05-29 01:35:00,893:DEBUG:i-0d2ea76e3e48e65d8:django_ses:[none]: send_messages.sent from=…

It also seems like using `list(reader)` (to see what the real, ungrouped
messages look like) before the `itertools.groupby` also fixes it, or at
least it prevented me from triggering it.

I could not reproduce it using the journal data itself and running it
locally, i.e. it might be related to the Python version (3.6 locally,
3.5 in the image), or more likely to some timing issues (since they are
available instantly when reading them locally).

Maybe there is even some special case / bug, because of the `else` being used
with `for` in this context.